requirements.txt
tensorflow
matplotlib
PyYAML
pandas
numpy
